# Improve performances of BigDecimal.numberOfDecimalDigits().