@CLAassistant 2024-06-10
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/ionspin/kotlin-multiplatform-bignum?pullRequest=299) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/ionspin/kotlin-multiplatform-bignum?pullRequest=299) before we can accept your contribution.<br/><hr/>**Grégory Lureau** seems not to be a GitHub user. You need a GitHub account to be able to sign the CLA. If you have already a GitHub account, please [add the email address used for this commit to your account](https://help.github.com/articles/why-are-my-commits-linked-to-the-wrong-user/#commits-are-not-linked-to-any-user).<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/ionspin/kotlin-multiplatform-bignum?pullRequest=299) it.</sub>