@glureau 2024-06-10
Thank you! It would be wonderful if I can get a release with that change soon. 🙏🏻 
Also I'm a bit confused, I've signed it but the bot still say it's not signed :(

<img width="1190" alt="Screenshot 2024-06-10 at 22 29 16" src="https://github.com/ionspin/kotlin-multiplatform-bignum/assets/7968075/42ca07a8-acb4-40ee-9ee7-d39e112ef1b8">