@ionspin 2024-06-10
I think the actual commit has to have the same email address you used to sign up with GitHub, I haven't read this article but I feel it's roughly in the direction of the solution :) https://docs.github.com/en/pull-requests/committing-changes-to-your-project/troubleshooting-commits/why-are-my-commits-linked-to-the-wrong-user