@ionspin 2024-06-10
The CI requires manual approval for pull requests, so no worries I'll run that. As far as replacing `== ZERO` with `isZero` to be honest I can't answer right now, I'd have to go through the source and remind myself for that particular case. If you want to submit a pull request with that change, I think that would help have an overview of where it was used.

Also you have an extra bracket in the test that makes it not compile, I'll leave a comment to point it out.