@ionspin 2024-06-10
Merged, thanks once more! Once the CI is complete it will publish new snapshot version